Isolate react components from react loaders
