+++ app/javascript/loaders/TypesDeChampEditor.js
@@ -1,0 +1,3 @@
+import Loadable from '../components/Loadable';
+
+export default Loadable(() => import('../components/TypesDeChampEditor'));
